+++ src/parsing/check_line_utils.c
@@ -74,7 +74,7 @@ int	check_double(char **doub_str, double lower_lim, double upper_lim)
 		(*doub_str)++;
 	}
 	str[++i] = '\0';
-	ratio = ft_atof(str);
+	ratio = (double)ft_atof(str);
 	if (ratio < lower_lim || ratio > upper_lim)
 		return (1);
 	return (0);
